refactor: update skeleton
