+++ src/index.ts
@@ -1,7 +1,11 @@
-import { validateSkippedStacks } from '@mnrendra/read-stacked-json'
-
 import mainAsync from './async'
 import mainSync from './sync'
+
+export {
+  type SkippedStacks,
+  type ValidSkippedStacks,
+  validateSkippedStacks
+} from '@mnrendra/read-stacked-json'
 
 export type {
   // @mnrendra/types-tsconfig
@@ -21,14 +25,10 @@ export type {
   WatchOptions,
   TypeAcquisition,
   // @mnrendra/read-tsconfig
-  Options,
-  // @mnrendra/validate-skipped-stacks
-  SkippedStacks,
-  ValidSkippedStacks
+  Options
 } from './types'
 
 export {
   mainAsync as readTSConfig,
-  mainSync as readTSConfigSync,
-  validateSkippedStacks
+  mainSync as readTSConfigSync
 }
