--- tests/index.ts
@@ -1,7 +1,0 @@
-import * as mocks from './mocks'
-import * as unmocks from './unmocks'
-
-export {
-  mocks,
-  unmocks
-}
