--- tests/mocks/index.ts
@@ -1,9 +1,0 @@
-import mockedReadAsync from './readAsync'
-import mockedReadSync from './readSync'
-import mockedStackTrace from './stackTrace'
-
-export {
-  mockedReadAsync,
-  mockedReadSync,
-  mockedStackTrace
-}
