--- tests/unmocks/index.ts
@@ -1,9 +1,0 @@
-import unmockReadAsync from './readAsync'
-import unmockReadSync from './readSync'
-import unmockStackTrace from './stackTrace'
-
-export {
-  unmockReadAsync,
-  unmockReadSync,
-  unmockStackTrace
-}
